backend/requirements.txt
# FastAPI Core
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy[asyncio]
asyncpg
alembic

# Utilities
pydantic
pydantic-settings
python-dotenv
email-validator

# Security
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]

google-generativeai
pypdf
langchain-text-splitters
